change printer to writer
