+++ sdk/test_data/writer/expected_object.yaml
