+++ sdk/test_data/writer/expected_objects.yaml
