+++ sdk/writer.go
@@ -10,17 +10,29 @@ import (
 	"github.com/nobl9/nobl9-go/manifest"
 )
 
+// WriteObjects writes objects to the given [io.Writer] in specified [manifest.ObjectFormat].
+func WriteObjects(objects []manifest.Object, out io.Writer, format manifest.ObjectFormat) error {
+	return writeObjects(objects, out, format)
+}
+
+// WriteObject writes a single object to the given [io.Writer] in specified [manifest.ObjectFormat].
+func WriteObject(object manifest.Object, out io.Writer, format manifest.ObjectFormat) error {
+	return writeObjects(object, out, format)
+}
+
 // PrintObjects prints objects to the given [io.Writer] in specified [manifest.ObjectFormat].
+// Deprecated: Use WriteObjects instead.
 func PrintObjects(objects []manifest.Object, out io.Writer, format manifest.ObjectFormat) error {
-	return printObjects(objects, out, format)
+	return writeObjects(objects, out, format)
 }
 
 // PrintObject prints a single object to the given [io.Writer] in specified [manifest.ObjectFormat].
+// Deprecated: Use WriteObject instead.
 func PrintObject(object manifest.Object, out io.Writer, format manifest.ObjectFormat) error {
-	return printObjects(object, out, format)
+	return writeObjects(object, out, format)
 }
 
-func printObjects(objects any, out io.Writer, format manifest.ObjectFormat) error {
+func writeObjects(objects any, out io.Writer, format manifest.ObjectFormat) error {
 	switch format {
 	case manifest.ObjectFormatJSON:
 		enc := json.NewEncoder(out)
